+++ src/main/java/com/bootcamp/savemypodo/controller/AuthController.java
@@ -5,6 +5,8 @@ import com.bootcamp.savemypodo.entity.User;
 import com.bootcamp.savemypodo.global.exception.ErrorCode;
 import com.bootcamp.savemypodo.global.exception.UserException;
 import com.bootcamp.savemypodo.repository.UserRepository;
+import com.bootcamp.savemypodo.service.AuthService;
+import com.bootcamp.savemypodo.service.UserService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -21,48 +23,14 @@ import java.util.Arrays;
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final JwtTokenProvider jwtTokenProvider;
-    private final UserRepository userRepository;
+    private final AuthService authService;
 
     @PostMapping("/refresh")
     public ResponseEntity<Void> refreshToken(HttpServletRequest request, HttpServletResponse response) {
-        String refreshToken = getTokenFromCookie(request, "refreshToken");
 
-        if (refreshToken == null) {
-            throw new UserException(ErrorCode.REFRESH_TOKEN_NOT_FOUND);
-        }
-
-        if (!jwtTokenProvider.validateToken(refreshToken)) {
-            throw new UserException(ErrorCode.REFRESH_TOKEN_EXPIRED);
-        }
-
-        String email = jwtTokenProvider.getEmailFromToken(refreshToken);
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new UserException(ErrorCode.USER_NOT_FOUND));
-
-        if (!refreshToken.equals(user.getRefreshToken())) {
-            throw new UserException(ErrorCode.REFRESH_TOKEN_MISMATCH);
-        }
-
-        // 🔄 새 AccessToken 발급
-        String newAccessToken = jwtTokenProvider.createAccessToken(user);
-
-        Cookie accessTokenCookie = new Cookie("accessToken", newAccessToken);
-        accessTokenCookie.setHttpOnly(true);
-        accessTokenCookie.setPath("/");
-        accessTokenCookie.setMaxAge((int) (jwtTokenProvider.getAccessTokenValidity() / 1000));
+        Cookie accessTokenCookie = authService.generateRefreshToken(request);
         response.addCookie(accessTokenCookie);
 
         return ResponseEntity.ok().build();
-    }
-
-    private String getTokenFromCookie(HttpServletRequest request, String name) {
-        if (request.getCookies() == null) return null;
-
-        return Arrays.stream(request.getCookies())
-                .filter(c -> c.getName().equals(name))
-                .findFirst()
-                .map(Cookie::getValue)
-                .orElse(null);
     }
 }
